Updated pipeline for adjusting the right location for the permissions (#5)
This pull request makes a small but important update to the CI/CD
workflow configuration. It adds explicit permissions for writing checks
and pull request updates, which is necessary for certain GitHub Actions
to function correctly.

- Workflow permissions:
*
[`.github/workflows/netcore-cicd.yml`](diffhunk://#diff-7c504959ef4fd3ab846d350c1842f9c6bbbdd92a51a061338f5e82741f7f7a8aR24-R27):
Added `checks: write` and `pull-requests: write` permissions to the
workflow to enable actions that require updating checks and pull
requests.